requirements.txt
argcomplete==1.9.3
astroid==2.0.4
base58==1.0.0
CacheControl==0.12.5
cachetools==2.1.0
certifi==2018.8.13
cfn-flip==1.0.3
chardet==3.0.4
click==6.7
docutils==0.14
durationpy==0.5
firebase-admin==2.12.0
Flask==1.0.2
Flask-Cors==3.0.6
future==0.16.0
grpcio==1.14.1
gunicorn==19.9.0
hjson==3.0.1
idna==2.7
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
jmespath==0.9.3
kappa==0.6.0
lambda-packages==0.20.0
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
msgpack==0.5.6
placebo==0.8.1
protobuf==3.6.1
psycopg2==2.7.5
pyasn1==0.4.4
pyasn1-modules==0.2.2
PyJWT==1.6.4
pylint==2.1.1
python-dateutil==2.6.1
python-slugify==1.2.4
pytz==2018.5
PyYAML==3.12
requests==2.19.1
rsa==3.4.2
s3transfer==0.1.13
six==1.11.0
toml==0.9.4
tqdm==4.19.1
troposphere==2.3.1
typed-ast==1.1.0
Unidecode==1.0.22
urllib3==1.23
Werkzeug==0.14.1
wrapt==1.10.11
wsgi-request-logger==0.4.6
